PROJECTS/yt-transcript-analyzer/requirements.txt
yt-dlp
webvtt-py
requests
python-dotenv
